requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==4.2.16
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
idna==3.6
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.6
django-environ==0.12.0
numpy==1.26.4
pandas==2.2.2
randomcolor==0.4.4.6

python-dateutil==2.9.0.post0
setuptools==74.1.2
stripe==11.1.0
tzdata==2024.1
xlrd==2.0.1